Burning captions into the video itself
This change allows us to burn the captions into the video itself, allowing for accurate measurements of captions to appear on the video without needing controls for it. Not entirely sure if this is the right longterm decision, but it is the one that I am leading with as of right now.