Got rid of redundant MediaConvert code.
I realized my previous code for attaching captions was working. I did all the extra work for nothing, so I deleted the extra code I added. Still need to do some error handling to ensure videos of the same title are allowed to be uploaded and captions can exist regardless. Transcribe does not like having multiple job names of the same name, so this is not yet finished.